[REFACTORING] Split IMAPServerIdleTest
